+++ q_compress/changelog.md
@@ -4,6 +4,8 @@
 
 * Added `use_gcds` compressor config and flag to improve compression ratio
 in cases where all numbers in some ranges share a nontrivial common divisor.
+* Made `CompressorConfig`, `DecompressorConfig`, `Prefix`, `Flags`, and
+`ChunkMetadata` more API-stable by adding a phantom private field.
 
 ## 0.9.3